Post-processes statuses composed in modals
